pyproject.toml
[project]
name = "cs-alignment-llms"
version = "0.1.0"
description = ""
authors = [
    {name = "Nolan Welch",email = "nolanwelch@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai (>=1.86.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
